# openapi/cure-agent.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "description": "소셜 콜백이 발급한 티켓과 한의원명·면허번호를 받아 가입을 마친다 (docs/specs/17). 이메일·소셜 신원은 티켓에서 꺼내므로 바디로 받지 않는다.",
        "operationId": "AuthController_completeSignUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteSignUpRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthSessionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "온보딩 완료 + 즉시 로그인 (쿠키 발급)",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cure Agent API",
    "description": "한의 임상 지침 기반 어시스턴트 API. 일반 JSON 응답은 공통 봉투(success/code/message/data/page/timestamp/traceId)를 사용하며, SSE 스트리밍에는 봉투를 적용하지 않는다. 인증은 HttpOnly 쿠키 기반이다.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "size": {
            "type": "number",
            "description": "요청한 페이지 크기 에코 (§10.4 — 항목 수는 data.length)"
          },
          "hasNext": {
            "type": "boolean",
            "description": "다음 페이지 존재 여부"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "다음 페이지 커서(불투명 문자열)"
          }
        },
        "required": [
          "size",
          "hasNext",
          "nextCursor"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "비즈니스 성공 여부"
          },
          "code": {
            "type": "string",
            "description": "코드 레지스트리 기반 식별자 (FE 분기용)"
          },
          "message": {
            "type": "string",
            "description": "사용자 표시용 메시지"
          },
          "page": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageMetaDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "응답 생성 시각 (ISO 8601)"
          },
          "traceId": {
            "type": "string",
            "description": "요청 추적 ID (로그·알림과 연결)"
          }
        },
        "required": [
          "success",
          "code",
          "message",
          "page",
          "timestamp",
          "traceId"
        ]
      },
...
      "ClinicSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ClinicianResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/ClinicSummaryResponseDto"
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "clinic",
          "verificationStatus"
        ]
      },
      "AuthSessionResponseDto": {
        "type": "object",
        "properties": {
          "clinician": {
            "$ref": "#/components/schemas/ClinicianResponseDto"
          },
          "expiresAt": {
            "type": "string",
            "description": "access 토큰 만료 시각 (ISO 8601)"
          }
        },
        "required": [
          "clinician",
          "expiresAt"
        ]
      },
      "CompleteSignUpRequestDto": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "string",
            "description": "소셜 콜백이 발급한 1회용 티켓"
          },
          "displayName": {
            "type": "string",
            "example": "김의사",
            "description": "기본값은 소셜 프로필 이름"
          },
          "clinicName": {
            "type": "string",
            "example": "서울한의원"
          },
          "licenseNumber": {
            "type": "string",
            "description": "면허번호 — 저장 시 암호화된다"
          },
          "termsAccepted": {
            "type": "boolean",
            "description": "true 필수"
          }
        },
        "required": [
          "ticket",
          "displayName",
          "clinicName",
          "licenseNumber",
          "termsAccepted"
        ]
      },
...
    }
  }
}